+++ src/mod_queue.rs
@@ -4,13 +4,7 @@ use std::sync::{Arc, Mutex};
 
 type Closure = Box<dyn FnOnce(&mut World, &mut Resources) + Send + Sync + 'static>;
 
-type ModQueuePoolArc = Arc<Mutex<Vec<Option<ModQueueInner>>>>;
-
-#[derive(Default)]
-struct ModQueueInner {
-    closures: Vec<Closure>,
-    dirty: bool,
-}
+type ModQueuePoolArc = Arc<Mutex<Vec<Option<Vec<Closure>>>>>;
 
 pub struct ModQueuePool {
     pool: ModQueuePoolArc,
@@ -38,23 +32,17 @@ impl ModQueuePool {
 
     pub fn new_mod_queue(&self) -> ModQueue {
         let mut pool = self.pool.lock().expect("mutexes should never be poisoned");
-        let (index, inner) = pool
+        let (index, closures) = pool
             .iter_mut()
             .enumerate()
-            .find(|(_, option)| {
-                if let Some(inner) = option {
-                    !inner.dirty
-                } else {
-                    false
-                }
-            })
+            .find(|(_, option)| option.is_some())
             .map(|(index, option)| (index, option.take()))
             .unwrap_or_else(|| {
                 pool.push(None);
                 (pool.len() - 1, Some(Default::default()))
             });
         ModQueue {
-            inner,
+            closures,
             index,
             pool: self.pool.clone(),
         }
@@ -63,19 +51,17 @@ impl ModQueuePool {
     pub fn apply_all(&self, world: &mut World, resources: &mut Resources) {
         let mut pool = self.pool.lock().expect("mutexes should never be poisoned");
         pool.iter_mut().for_each(|option| {
-            if let Some(ref mut inner) = option {
-                inner
-                    .closures
+            if let Some(ref mut closures) = option {
+                closures
                     .drain(..)
                     .for_each(|closure| closure(world, resources));
-                inner.dirty = false;
             }
         });
     }
 }
 
 pub struct ModQueue {
-    inner: Option<ModQueueInner>,
+    closures: Option<Vec<Closure>>,
     index: usize,
     pool: ModQueuePoolArc,
 }
@@ -83,18 +69,14 @@ pub struct ModQueue {
 impl Drop for ModQueue {
     fn drop(&mut self) {
         let mut pool = self.pool.lock().expect("mutexes should never be poisoned");
-        pool[self.index] = self.inner.take();
+        pool[self.index] = self.closures.take();
     }
 }
 
 impl ModQueue {
-    fn get_inner_mut(&mut self, will_be_dirty: bool) -> &mut Vec<Closure> {
-        self.inner
+    fn closures_mut(&mut self) -> &mut Vec<Closure> {
+        self.closures
             .as_mut()
-            .map(|inner| {
-                inner.dirty = will_be_dirty;
-                &mut inner.closures
-            })
             .expect("modification queue should contain the inner queue at this point")
     }
 
@@ -102,6 +84,6 @@ impl ModQueue {
     where
         F: FnOnce(&mut World, &mut Resources) + Send + Sync + 'static,
     {
-        self.get_inner_mut(true).push(Box::new(closure))
+        self.closures_mut().push(Box::new(closure))
     }
 }